This PR introduces the following optimisations. 

# Major Changes

1. Rewrite _draw_random_discreet_gaussian. Shaved 30% of the run time when I ran it on my (now outdated) branch (159 sec --> 104 sec). 

2. Commented out `assert_state_changes`. It saves 12% run time (91 sec --> 81 sec). 

3. Sample multiple random variables at once in `Human.symptoms` (81 sec --> ~75 sec). 

# Minor/Cosmetic Changes

* Reformated utils.py. It was mixing tabs with spaces (we should really try to avoid that).